add test cases against the cli command
