+++ test/jasmine.json
@@ -2,7 +2,8 @@
   "spec_dir": "test",
   "spec_files": [
     "spec_lib/*.js",
-    "spec_rule/*.js"
+    "spec_rule/*.js",
+    "spec_cli/*.js"
   ],
   "helpers": [
      "../node_modules/babel-register/lib/node.js",
